# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ATLAS Kim \u2014 Unified RS Intelligence Engine",
    "description": "Deterministic relative-strength intelligence layer for equities, MFs, ETFs and global indices.",
    "version": "1.0.0"
  },
  "paths": {
    "/api/unified/snapshot/{instrument_id}": {
      "get": {
        "tags": [
          "unified"
        ],
        "summary": "Instrument evidence card",
        "description": "Returns identity, latest metrics, lookthrough, factor percentiles, and deterministic narrative.",
        "operationId": "get_snapshot_api_unified_snapshot__instrument_id__get",
        "parameters": [
          {
            "name": "instrument_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Instrument Id"
            }
          },
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional snapshot date (defaults to latest)",
              "title": "Date"
            },
            "description": "Optional snapshot date (defaults to latest)"
          },
          {
            "name": "tenant_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Tenant identifier",
              "default": "default",
              "title": "Tenant Id"
            },
            "description": "Tenant identifier"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/unified/aggregate": {
      "get": {
        "tags": [
          "unified"
        ],
        "summary": "Cohort aggregate (bubble + table)",
        "description": "Returns aggregated metrics grouped by cohort (sector, cap, state, mf_category, country, global_sector, etc.).",
        "operationId": "get_aggregate_api_unified_aggregate_get",
        "parameters": [
          {
            "name": "cohort_type",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CohortType",
              "description": "Dimension to group by"
            },
            "description": "Dimension to group by"
          },
          {
            "name": "benchmark",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Benchmark",
              "description": "Benchmark for RS calculation",
              "default": "nifty"
            },
            "description": "Benchmark for RS calculation"
          },
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Period",
              "description": "Return period for median",
              "default": "3m"
            },
            "description": "Return period for median"
          },
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional snapshot date (defaults to latest)",
              "title": "Date"
            },
            "description": "Optional snapshot date (defaults to latest)"
          },
          {
            "name": "instrument_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by instrument type (e.g. EQUITY, MF, ETF, INDEX_GLOBAL)",
              "title": "Instrument Type"
            },
            "description": "Filter by instrument type (e.g. EQUITY, MF, ETF, INDEX_GLOBAL)"
          },
          {
            "name": "tenant_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Tenant identifier",
              "default": "default",
              "title": "Tenant Id"
            },
            "description": "Tenant identifier"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AggregateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/unified/screen": {
      "post": {
        "tags": [
          "unified"
        ],
        "summary": "Filterable screener",
        "description": "Apply metric filters and return paginated rows.",
        "operationId": "post_screen_api_unified_screen_post",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Tenant identifier",
              "default": "default",
              "title": "Tenant Id"
            },
            "description": "Tenant identifier"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScreenerRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScreenerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/unified/regime": {
      "get": {
        "tags": [
          "unified"
        ],
        "summary": "Current market regime",
        "description": "Returns the latest unified market regime with deterministic narrative.",
        "operationId": "get_regime_api_unified_regime_get",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional regime date (defaults to latest)",
              "title": "Date"
            },
            "description": "Optional regime date (defaults to latest)"
          },
          {
            "name": "tenant_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Tenant identifier",
              "default": "default",
              "title": "Tenant Id"
            },
            "description": "Tenant identifier"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegimeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/unified/funds/screen": {
      "post": {
        "tags": [
          "unified"
        ],
        "summary": "Filter by factors, category, cap tilt",
        "description": "Screen funds using 6-factor percentiles and category filters.",
        "operationId": "post_fund_screen_api_unified_funds_screen_post",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Tenant identifier",
              "default": "default",
              "title": "Tenant Id"
            },
            "description": "Tenant identifier"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FundScreenRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FundScreenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CohortType": {
        "type": "string",
        "enum": [
          "sector",
          "cap_category",
          "mf_category",
          "instrument_type",
          "state",
          "action",
          "country",
          "global_sector"
        ],
        "title": "CohortType"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SnapshotResponse": {
        "properties": {
          "instrument": {
            "$ref": "#/components/schemas/InstrumentIdentity"
          },
          "metrics": {
            "$ref": "#/components/schemas/MetricSnapshot"
          },
          "lookthrough": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LookthroughSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "factor_percentiles": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FactorPercentiles"
              },
              {
                "type": "null"
              }
            ]
          },
          "narrative": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NarrativeBlock"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "type": "object",
        "required": [
          "instrument",
          "metrics",
          "meta"
        ],
        "title": "SnapshotResponse"
      },
...
    }
  }
}